refactor(player): move window drag service into core/window
Makes room for window geometry persistence, which needs the same
windowManager.ensureInitialized() call. Two modules racing to initialize
the same singleton would be a bug, so they share one home.

Also adds tool/host_flutter.sh: on a devenv-loaded macOS worktree the
Nix-wrapped flutter cannot build native assets, so the suite needs the
same escape hatch `./dev player macos` already uses.
