+++ player/lib/core/window/window_drag_service_stub.dart
